LayoutTests/media/media-source/media-source-append-nonsync-sample-after-abort-expected.txt


RUN(video.src = URL.createObjectURL(source))
RUN(quality = video.getVideoPlaybackQuality())
EXPECTED (quality.droppedVideoFrames == '0') OK
EXPECTED (video.duration == '4') OK
RUN(quality = video.getVideoPlaybackQuality())
EXPECTED (quality.droppedVideoFrames == '5') OK
END OF TEST
